More work on Remote Monitor interface
git-svn-id: http://svn.code.sf.net/p/perfmon4j/code/trunk@194 38ac2094-77c1-42ba-995b-c18f1c0364f1

+++ base/src/main/java/org/perfmon4j/remotemanagement/SessionManager.java
@@ -38,7 +38,7 @@ class SessionManager<T extends SessionManager.SessionData> {
 	static final int NO_TIMEOUT = -1;
 	static private final int DEFAULT_REAPER_MILLIS = 
 		Integer.getInteger(SessionManager.class.getName() + ".DEFAULT_REAPER_MILLIS",
-				10 * 60 * 1000); // Ten minutes by default.
+				10 * 60 * 1000).intValue(); // Ten minutes by default.
 
 	private final Object mapLockToken = new Object();
 	private final Map<String, SessionWrapper> map = new HashMap<String, SessionWrapper>();
